pyproject.toml
[tool.poetry]
name = "checkin-tool"
version = "0.1.0"
description = ""
authors = ["Thanh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
